Pulling out sendable implementations into their own python modules
