+++ fa_search_bot/tests/integration/sites/e621/test_e621_post.py
@@ -1,6 +1,6 @@
 import pytest
 
-from fa_search_bot.sites.e621.e621_handler import E621Post
+from fa_search_bot.sites.e621.sendable import E621Post
 
 
 @pytest.mark.asyncio
